+++ src/components/workflow/panels/StoreServiceRiskPanel.jsx
@@ -16,6 +16,8 @@ import * as NRS from '../../../data/networkRiskSignals'
 import * as SCB from '../../../data/scenarioBStranding'
 import * as SHC from '../../../data/scenarioHubClosure'
 import * as SCP from '../../../data/scenarioCrossHub'
+import * as SMX from '../../../data/scenarioMisconnect'
+import * as STC from '../../../data/scenarioTailCrew'
 import * as SPG from '../../../data/scenarioPolicyGate'
 import { NetworkMap, BeforeAfterFlow, EfficientFrontier, WorkflowActions, ModelUpdateGrid } from '../../viz/ScmViz'
 
@@ -28,6 +30,8 @@ const MODULES = {
   'uc-mid-rotation-stranding': { prefix: 'SCB', mod: SCB },
   'uc-hub-closure': { prefix: 'SHC', mod: SHC },
   'uc-cross-hub-propagation': { prefix: 'SCP', mod: SCP },
+  'uc-misconnect-exposure': { prefix: 'SMX', mod: SMX },
+  'uc-tail-crew-sync': { prefix: 'STC', mod: STC },
   'uc-policy-feasibility-gate': { prefix: 'SPG', mod: SPG },
 }
 function useData() {